Moved code to inc/Engine/Cache/Purge.php
